+++ utils/users.js
@@ -1,36 +1,15 @@
-import _ from "lodash"
-
-export const getUrisOfUser = user => {
-  if (!user) {
-    return []
+export function expandIdentities(identities, identityGroups) {
+  if (identities && identityGroups) {
+    return identities.map(uri => uri in identityGroups ? identityGroups[uri].identities : uri).flat()
   }
-  return [user.uri].concat(Object.values(user.identities || {})
-    .map(identity => identity.uri)).filter(uri => uri)
+  return identities
 }
 
-/**
- * Extracts a creator objects from a request.
- *
- * @param {*} req request object
- */
-export const getCreator = (req) => {
-  const user = getUser(req)
+export const getUrisOfUser = user =>
+  [user || {}, ...Object.values(user?.identities || {})]
+    .map(({uri}) => uri).filter(Boolean)
 
-  const creator = {}
-
-  if (user.uri) {
-    const creatorUriPath = req.type === "annotations" ? "id" : "uri"
-    _.set(creator, creatorUriPath, user.uri)
-  }
-  if (user.name) {
-    const creatorNamePath = req.type === "annotations" ? "name" : "prefLabel.en"
-    _.set(creator, creatorNamePath, user.name)
-  }
-
-  return user.uri || user.name ? creator : null
-}
-
-export const getUser = (req) => {
+export const getUser = req => {
   const user = {}
 
   const userUris = getUrisOfUser(req.user)
@@ -43,11 +22,32 @@ export const getUser = (req) => {
   if (req.query.identityName) {
     user.name = req.query.identityName
   } else if (req.query.identityName !== "") {
-    const name = _.get(Object.values(_.get(req, "user.identities", [])).find(i => i.uri === user.uri) || req.user, "name")
-    if (name) {
-      user.name = name
+    const identity = Object.values(req?.user?.identities || {}).find(({uri}) => uri === user.uri) || req.user
+    if (identity?.name) {
+      user.name = identity.name
     }
   }
 
   return user
 }
+
+export const getCreator = req => {
+  const { uri, name } = getUser(req)
+  const creator = {}
+
+  if (uri) {
+    creator[req.type === "annotations" ? "id" : "uri"] = uri
+  }
+  if (name) {
+    if (req.type === "annotations") {
+      creator.name = name
+    } else {
+      creator.prefLabel ||= {}
+      creator.prefLabel.en = name // TODO: why en?
+    }
+  }
+
+  return uri || name ? creator : null
+}
+
+